introduce the design of two new tables
This adds the database schemas for two new tables. One for email addresses
(EMAIL_ADDRESS) and another for phone numbers (PHONE_NUMBER).

The reason for this is that nowadays it is pretty common for someone to have
more than one email address and more than one phone number. For example,
one personal email and another for work. Same thing with phone numbers, one for
home (personal), one mobile and another for business (work).

So we want to allow the user to be able to store a list of emails and phones
of his contacts.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>