learn/c: demonstrate how to allocate memory for an array on the heap
This examples shows how you can use malloc (or calloc) to allocate memory for
an array on the heap and also the relationship between pointers and arrays.

It also demonstrates that an array is really a contiguously allocated
nonempty set of objects of a particular type (for example, in our case, the
type is int), by the multiplication of the numbers of elements by the size of
each element.

So, say you need an array of 5 slots, the following is what it will look like
in memory:

    	[][][][][]
index	0 1 2 3 4

Refs:
https://learn.microsoft.com/en-us/cpp/cpp/arrays-cpp?view=msvc-170
https://www.embedded.com/allocating-arrays/
http://cs.ecu.edu/karl/3300/spr16/Notes/C/Array/heap.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>