tribe: store the number of rows deleted by the DELETE statement
sqlite3_changes() return the numbers of rows deleted by the most recently
completed DELETE statement.

We want to know that value to check if the DELETE statement actually deleted
something or not.

If you run the DELETE query on an id that does not exist on the table, changes
should report 0.

We are reserving this for future use, because it will need to provide to the
caller a status on the query executed, so it can check if something was deleted
or not.

Refs:
https://www.sqlite.org/c3ref/changes.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>