tribe: delete contacts from the command line without loading all the contacts
Why? Because, it should not be necessary to load on to memory all the
contacts from the database just to delete a single entry. If the database
is huge, it will require a lot of memory, and will waste a bunch of time.
In the worst case, given a non-existent contact ID, it will iterate
through all the contacts to not find it.

By taking the issue directly to the database and executing the DELETE query
right way, we spend less time and save a bunch of memory.

And as a bonus, it will allow us to remove the ab_get_contact_by_id function
in the future. Which will help us with our goal of making the ab api as
stateless as possible.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>