tribe: change the implementation of ab_get_contact_by_id function
Change it to talk directly to the database rather than iterating
through the linked list.

The old way, by iterating through the linked list, required it
to load all the contacts from the database before by calling
ab_load_contacts function. I didn't like this, because if
there was huge number of contacts saved in the database,
loading them could require a lot of memory and waste a bunch
of precious resources and time.

By talking directly to the database, loading the contact list
is not required anymore.

That also helps with our goal of making the api stateless as possible,
and moving all the interaction with the linked list into the client
code.

Did a bunch of tests, and they seem to be working correctly as before.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>