numbers: add second version to show the use of double pointers
It shows some techniques, but the most important is the pointer to pointer (
double pointer) as an output parameter.

If also shows that if you take the address of a pointer variable that creates
a double indirection (i.e., a pointer to pointer variable).

Another interesting thing to note is that it also demonstrates how a pointer
to pointer (double pointer) can be used to return to the function caller, a
pointer to the beginning of space (the memory location) allocated for the
array.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>