tribe: use goto to properly deallocate memory in add command
I did not find many articles on this topic. The only good references
I found on this were [1] and [2].

Using return statements in the middle of the code leds to too much
code duplication. You have to do the same thing, over and over. And
as it was happening with our code, we were never closing the database,
for example, if some error occurred, because we were returning early.
With the goto statement, you define a single area where you free, close,
and release the objects and resources previously used.

Refs:
[1] - https://eli.thegreenplace.net/2009/04/27/using-goto-for-error-handling-in-c
[2] - https://vilimpoc.org/research/raii-in-c/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>