tribe: change the standalone commands to have shorter names
One of the goals was to remove "tribe" from the filenames. The other was
to not have hyphens ("-") in the program name. All that was possible when
I realized I could build the program names from CMake. By not having the
program name on the files will make it much easier to rename the project (again)
in the future if I decided to do so.

The end result was:

tribecmd-add -> tribeadd
tribecmd-mod -> tribemod
tribecmd-del -> tribedel
tribecmd-ls  -> tribels

Signed-off-by: Thiago Farina <tfransosi@gmail.com>