tribe: rename ab_load_contacts function to ab_get_contacts
Now that we renamed _db_get_contact_list function in commit b57fb8c ("tribe:
rename _db_get_contact_list function to _db_get_contacts") to _db_get_contacts,
it makes more sense to also rename ab_load_contacts function to
ab_get_contacts (specially the v2, as it is just a wrapper around
_db_get_contacts).

Another thing that motivated this and prompted me to do it was the use of
"load" verb in this function name. The "load" verb is not associated with
collections, list of items or the enumaration of items. It is more commonly
associated (and thus used) with the "load" of a resource or a module.
Let's look at some functions from the Windows API winuser.h and use them as a
good example of properly naming functions.
Most of the LoadXXX functions are not used for enumerating a collection of items. Take LoadIcon() for example, it "loads" the specified icon RESOURCE
from the executable (.exe) file. The same thing for LoadCursor, LoadImage,
LoadMenu and LoadString. They all operate on a SINGLE item and usually return
a HANDLE as its return value and not an array by the means of a pointer to
pointer.

Refs:
https://learn.microsoft.com/en-us/windows/win32/api/winuser/nf-winuser-loadicona
https://gist.github.com/maxtruxa/b2ca551e42d3aead2b3d

Signed-off-by: Thiago Farina <tfransosi@gmail.com>