tribe: change the parameter of ab_contact_alloc function to double pointer
That way we can return the address of the space allocated for the ab_contact_t
structure, as a pointer to pointer stores the address of another pointer
varible. That technique allow us to return a status error code
from ab_contact_alloc function, like most (all) functions in ab.h.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>