tribe: copy the contacts array entries over to the linked list
I discovered this issue when I added a FREELIST(contact_list); to ab_fini
function after adding a call to ab_fini at commit 82b68de ("tribe: add a call
to ab_fini function before the application exits").

When I tried to free the contact_list, the application crashed. This was
because, when trying to free the inner "data" members, it was actually
trying to free the memory allocated for the array, and that was not going
well after freeing the first element.

Because we were not duplicating the entries, I couldn't even delete the array
after populating the linked list, as the data was in the region allocated
for the array.

Now, with the entries duplicated, it should be possible to free the array after
populating the linked list and free the linked list and its inner elements when
destroying the application window.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>