tribe: create contact object before opening the properties dialog
This will allow us to have better control at what goes into the properties
dialog and will allow the port to Windows code easier.

There is a problem with memory allocation we will have to address later on
though. The issue is that each time the user wants to create a new contact,
a new contact object is created, but if he decides to cancel and not save it,
then an object was created but not "used" and never released. If he keeps
doing that, memory usage will keep increasing. So that is something we will
have to fix in the future somehow.

This is a follow up to commit 2f0401e ("tribe: prevent critical NULL assertion
from gtk_entry_set_text function") and what actually originated it.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>