tribe: simplify the design of the "New" button on the toolbar
The old way, a button with a menu was for the case where we would have two
more menu items ("New Group" and "New Folder"). Since I'm not planning to
add groups and folders any time soon, as they add way more complexity to the
code base, it makes more sense to have a flat "New" button. Note that the
Windows version already has a flat version of this button. So they both are now
in sync, making the interface between the two platforms more uniform.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>