tribe: duplicate the string value inside the ab_contact_set_xxx functions
This has the benefit of me not having to add string duplication code in all
these places that call the set_xxx functions. Specially in the command line
tools where I was passing argv pointers.

Moreover, it consolidates the ownership of these strings inside the contact
object. Making it possible to manage more effectively.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>