tribe: set return code to ENOMEM when malloc fails
Let's start using ENOMEM and EINVAL to differentiate the errors that might
occur and be returned by a function.

On my system ENOMEM is defined at /usr/include/asm-generic/errno-base.h. I used
the command "locate errno.h" to find the header files.

Their counterparts from HRESULT are E_OUTOFMEMORY and E_INVALIDARG that can
be found in winerror.h file.

This will help when porting this code to Windows.

Refs:
https://learn.microsoft.com/en-us/windows/win32/seccrypto/common-hresult-values

Signed-off-by: Thiago Farina <tfransosi@gmail.com>