c: move naming enumerations into a dedicated section
They are very close and very related to macros and have a additional set of
rules specific to them, so it is better to move it into a separate section.

NOTE: I have to expand the code examples to demonstrate how enum values can
be named (there are some variations and possibilities that can be used) and
initialized.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>