c: demonstrate why you need a macro indirection to stringize a macro's value
This will show that if you apply the preprocessing operator '#' directly to a
macro, it will just put the macro's name in quotes. To put the macro's value
in quotes, you to expand it first, then apply the operator '#' on it.

See the example stringize_test.c file to see the results.

I'm adding this example to show how it works, because I will need this macro
to build a version string out of the version numbers in the tribe project.

Refs:
https://c-faq.com/ansi/stringize.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>