buffer: define ARFLAGS variable and use it
This moves the 'rcs' flags we pass to 'ar' into a variable,
because we are going to change these flags in a future patch
and I want git blame to easily show what has changed. If it
remains in the same line, it is harder to spot what is being
changed.

Tested with:

$ make clean
$ make buffer_test

Signed-off-by: Thiago Farina <tfransosi@gmail.com>