buffer: declare the clean target to be phony
Without declaring it to be phony, the clean target would not work
properly if a file named clean is ever created in the current directory.
Since it has no prerequisites, clean would always be considered up to date
and its recipe would not be executed. To avoid this problem we have to declare
the it to be phony by making it a prerequisite of the special target .PHONY (see [2]).

For more in depth explanation refer to [1].

Refs:
[1] - https://www.gnu.org/software/make/manual/html_node/Phony-Targets.html
[2] - https://www.gnu.org/software/make/manual/html_node/Special-Targets.html
[3] - https://www.gnu.org/software/make/manual/html_node/Cleanup.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>