learn/assembly: change the extension of exit.asm to .s
The file name suffix (or extension) is important, because it
determines what kind of compilation gcc does.

For automake only the suffixes .s, .S, and .sx are recognized as
being files containing assembly code. [3]

For it, files with .s are assembler code, and files with .S
are assembler code that must be preprocessed by cpp first. [2]

See section `Options Controlling the Kind of Output' at [1].

Refs:
[1] - https://gcc.gnu.org/onlinedocs/gcc/Overall-Options.html
[2] - https://ftp.gnu.org/old-gnu/Manuals/gas-2.9.1/html_chapter/as_3.html#SEC26
[3] - https://www.gnu.org/software/automake/manual/html_node/Assembly-Support.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>