buffer: rename buffer_test.c to buffer_example.c
Because it is not really a test suite like vector_test.c is, as it does not use
the greatest.h test framework. It is just a small program that shows some usage
of the buffer API functions.

This opens the possibility to add another buffer_test.c which will use
make use of the greatest.h test framework to test the various buffer API
functions.

I'm also doing this because my plan is to merge buffer and vector under
a common directory.

Tested with make:

$ make clean
$ make
$ ./buffer_example

Tested with cmake:
$ mkdir out
$ cd out
$ cmake -GNinja ..
$ ninja -C out
$ ./out/buffer_example

Signed-off-by: Thiago Farina <tfransosi@gmail.com>