buffer: add test suite infrastructure
This adds all the files necessary to run a test suite for buffer. These
files were copied over from vector library and modified to build and run
for buffer.

This is a follow up to commit 476bc4a ("buffer: rename buffer_test.c to
buffer_example.c").

Tested with the following commands for make:

$ make buffer_unittests
$ ./buffer_unittests

Tested with the following commands for cmake:

$ ninja -C out buffer_unittests
$ ./out/buffer_unittests

Signed-off-by: Thiago Farina <tfransosi@gmail.com>