vector: remove examples and tests directories
This is because we want to have the same directory layout
for both buffer and vector, so in the next few patches
we can put them together under the same directory.

We decided that vector and buffer should be viewed as individual
modules (i.e., a pair of .h and .c) rather than individual libraries
(libxxx.a). Together they can be part of and form a new library, not be
individual libraries themselves.

I also plan to stop versoning it, because this will simplify the
maintainance and will be easier to put it together with buffer in
the same library.

Tested with the following commands with make:

$ make clean
$ make
$ ./vector_unittests
$ ./vector_example

Tested with the following commands with cmake:

$ ninja -C out vector_unittests vector_example
$ ./out/vector_unittests
$ ./out/vector_example
$ ninja -C out check

Signed-off-by: Thiago Farina <tfransosi@gmail.com>