tribe: always build with gtk
The main goal is to build the gtk application, so requiring
a switch does not make much sense and is error prone.

With the option, you had to pass WITH_GTK=ON to cmake like this:

$ cd build # or out
$ cmake -GNinja -DWITH_GTK=ON ..

Now you don't have, and the build instructions in the README file
are in sync with the CMake build file.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>