tribe: consolidate the sqlite3_finalize code in _db_update_contact function
This moves the sqlite3_finalize code into the clean up area (the out labeled
statement) to consolidate the two calls to sqlite3_finalize function.

This is to make sure that when the goto statements are executed in the error
path for the sqlite3_bind_xxx calls, the sqlite statement is properly
finalized, its return value is checked and an error message is printed if
necessary. The old code would not do that.

I also removed the NULL check for the sqlite statement, because calling
sqlite3_finalize() on a NULL pointer is harmless (i.e., it will not crash the
application).

Refs:
https://www.sqlite.org/c3ref/finalize.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>