tribe: consolidate the sqlite3_finalize code in _db_delete_contact function
This moves the sqlite3_finalize code into the clean up area (the out labeled
statement) to consolidate the two calls to sqlite3_finalize function.

This is to make sure that when the goto statements are executed in the error
handling paths, the sqlite statement is properly finalized, the statement is
assigned to NULL, its return value is checked and an error message is printed
if necessary. The old code was not doing all this in the labeled statement.

I also removed the NULL check for the sqlite statement, because calling
sqlite3_finalize() on a NULL pointer is harmless (i.e., it will not crash the
application).

This is a follow up to the work done at commit 297a436 ("tribe: consolidate
the sqlite3_finalize code in _db_update_contact function").

Refs:
https://www.sqlite.org/c3ref/finalize.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>