tribe: put sqlite3_finalize code in the out labeled statement
This moves the sqlite3_finalize code in the ab_get_contact_by_id
function to the out labeled statement. Why? Because this way,
it will be called if sqlite3_bind_int or sqlite3_step fail. The old
code was not calling the sqlite3_finalize function, if these functions
returned an error value.

Tested issuing tribemod and tribecli mod commands.

This is a follow up to the work done at commit 63ab3ec ("tribe:
consolidate the sqlite3_finalize code in _db_delete_contact function").

Signed-off-by: Thiago Farina <tfransosi@gmail.com>