tribe: check the return value of sqlite3_changes in _db_delete_contact
When trying to enable -Werror for source code under src/, it triggers
the following unused-but-set-variable error:

../src/ab.c: In function ‘_db_delete_contact’:
../src/ab.c:463:7: error: variable ‘rows_deleted’ set but not used [-Werror=unused-but-set-variable]
   int rows_deleted;

Checking the return value of sqlite3_changes function fixes this error.

We may print an error message and doing something else with it later though.

This is a continuation of the work started at commit 4702871 ("tribe: store
the number of rows deleted by the DELETE statement").

Signed-off-by: Thiago Farina <tfransosi@gmail.com>