hello: specify the minimum version of automake
If the version of automake installed on your system is not the same
or it is not newer than the 1.8 specified here, the creation of the
Makefile.in will be suppressed.

I picked an older version, because the one I have installed here is:

$ automake --version
automake (GNU automake) 1.15

Also note that AM_INIT_AUTOMAKE has a single argument, which is a space-separated
list of options that should be applied to every Makefile.am in the tree.

Tested with the following commands:

$ ./autogen.sh
$ ./configure
$ make

Refs:
https://www.gnu.org/software/automake/manual/html_node/Public-Macros.html
https://www.gnu.org/software/automake/manual/html_node/List-of-Automake-options.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>