acceptanddisplay.c: calculate the size of the array using sizeof
This is to avoid hardcoding constants, which is a bad practice.

If, for some reason, we change the size of the array buffer in the
first_name variable, we also need to change the fgets line, but we
might forget, so it is error prone. By calculating it with sizeof,
this problem is avoided.

It works fine as before.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>