tribe: rename sqlite library from sqlite3 to sqlite
The main reason is to avoid conflicts with the version
installed on the system, though that is very unlikely
to happen in this cmake build.

Before this change, the output would be libsqlite3.a.

Normally, the sqlite3 library is installed on a Ubuntu system
like this:

$ sudo apt install sqlite3

And with that, running pkg-config:

$ pkg-config --libs

Would return -lsqlite3.

Now, with this change, the output will be libsqlite.a. And that
will match with the directory name, which is a nice thing and another
reason for doing this.

To see the sqlite3's pkg-config in the fossil repository, follow the
refs link.

Refs:
https://www.sqlite.org/cgi/src/file?name=sqlite3.pc.in&ci=trunk

Signed-off-by: Thiago Farina <tfransosi@gmail.com>