tribe: add the word "manager" to the program description in the about dialog
The word "manager" is a noun, and is defined by Cambridge Dictionary [1] as
"a computer program that helps you organize a particular type of activity".

We are adding it to the description because it describes a program that helps
you organize your address book or contacts list.

Other example of apps that use "manager" to describe them are "archive manager",
"file manager", "password manager", "display manager", "user manager", etc.

Other common nouns used in other types of applications are "player" for media, video,
audio apps, "viewer" for document editing, "editor" for text, "viewer" for image,
"browser" for web, "reader" for mail, "emulator" for terminal, etc.

Refs:
[1] - https://dictionary.cambridge.org/dictionary/english/manager

Signed-off-by: Thiago Farina <tfransosi@gmail.com>