tribe: free contact's properties
If we don't free the existing string before duplicating the new string,
we lost the reference to it (that is the address that the pointer points to)
and leave that memory behind.

The issue is with calling set_xxx functions repeatedly like:

ab_contact_set_first_name(contact, "Name01");
ab_contact_set_first_name(contact, "Name02");
ab_contact_set_first_name(contact, "Name03");

In the first call fname might be initialized to NULL and that is fine,
then malloc is called and memory space is reserved for Name01.

What happens to fname in the second call? In the first call memory space
was already allocated for it and malloc returned the pointer to that area,
but now we are doing it again for Name02 and we were not freeing or releasing
the previous space reserved for Name01. That was leaving the area where
Name01 was at used, but changing the pointer of fname to another location,
to the location where Name02 is going to be. Hence, that is why we need
to free whatever fname is pointing to before allocating any more memory
for the new string.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>